Add Empire Node presentation to community events
I'm not sure whether the scope of this page is limited to talks given by people working on the standard. If it is not, this is a helpful talk that demonstrates using WASM with Node. The Node workflow is still clunky, but I think adventuresome developers might want to tinker with it, and this is the only demo of it I've seen.